将源相册放入temp，路径发送到mq，消费者将文件压缩为webp 读出元数据

+++ MyStep/MyStep_ZaoHu/MyStep-step/src/main/java/org/zaohu/common/entity/PhotoBatch.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Data
 public class PhotoBatch {
     private List<MultipartFile> photoTypeList;
-    private long typeId;
+    private Integer typeId;
 }
